+++ example/Main.hs
@@ -1,16 +1,22 @@
-{-# LANGUAGE DataKinds       #-}
-{-# LANGUAGE TypeOperators   #-}
+-- ConstraintKinds needed only for 7.8.4
+{-# LANGUAGE ConstraintKinds   #-}
+{-# LANGUAGE DataKinds         #-}
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TypeOperators     #-}
 module Main (main) where
 
-import Lucid
-import Lucid.Servant
-import Data.Maybe         (fromMaybe)
-import Network.Wai        (Application)
-import Servant
-import Servant.HTML.Lucid
-import System.Environment (getArgs, lookupEnv)
-import Text.Read          (readMaybe)
+import           Data.Maybe
+                 (fromMaybe)
+import           Lucid
+import           Lucid.Servant
+import           Network.Wai
+                 (Application)
+import           Servant
+import           Servant.HTML.Lucid
+import           System.Environment
+                 (getArgs, lookupEnv)
+import           Text.Read
+                 (readMaybe)
 
 import qualified Network.Wai.Handler.Warp as Warp
 
@@ -20,8 +26,13 @@ type API = "string" :> Get '[HTML] String
 api :: Proxy API
 api = Proxy
 
-stringLink :: Link
-stringLink = safeLink (Proxy :: Proxy API) (Proxy :: Proxy ("string" :> Get '[HTML] String))
+apiLink_
+    :: (IsElem endpoint API, HasLink endpoint)
+    => Proxy endpoint -> MkLink endpoint Attribute
+apiLink_ = safeAbsHref_ (Proxy :: Proxy API)
+
+stringLink_ :: Attribute
+stringLink_ = apiLink_ (Proxy :: Proxy ("string" :> Get '[HTML] String))
 
 server :: Server API
 server = return "example" :<|> return html where
@@ -29,7 +40,7 @@ server = return "example" :<|> return html where
     html = do
         p_ $ b_ "bar"
         p_ $ i_ "iik"
-        p_ $ a_ [absHref_ stringLink] "string"
+        p_ $ a_ [stringLink_] "string"
 
 app :: Application
 app = serve api server